다형성-실습(수정)

+++ include/TextureManager.h
@@ -20,10 +20,10 @@ public:
 
     bool load(string fileName, string id, SDL_Renderer* pRenderer);
 
-    void draw(string id, int x, int y, int width, int height,
+    void draw(string id, float x, float y, int width, int height,
         SDL_Renderer* pRenderer, SDL_RendererFlip flip = SDL_FLIP_NONE);
 
-    void drawFrame(string id, int x, int y, int width, int height,
+    void drawFrame(string id, float x, float y, int width, int height,
         int currentRow, int currentFrame,
         SDL_Renderer* pRenderer, SDL_RendererFlip flip = SDL_FLIP_NONE);
